# reference/lib/api-zod/src/generated/types/requestUploadUrlInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Professional Management Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface RequestUploadUrlInput {
  /** @minLength 1 */
  name: string;
  /** @minimum 0 */
  size: number;
  contentType: string;
}