# reference/lib/api-zod/src/generated/types/uploadUrlResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Professional Management Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { UploadUrlResultMetadata } from './uploadUrlResultMetadata';

export interface UploadUrlResult {
  uploadURL: string;
  objectPath: string;
  metadata: UploadUrlResultMetadata;
}